requirements.txt
selenium
beautifulsoup4
requests
pandas
PyPDF2
chromadb
psycopg2-binary
python-dotenv
zipfile36
pdfplumber
tqdm
nsepython
sqlalchemy
camelot-py[base]
matplotlib
llama-parse
llama-index
google-genai
ollama
numpy
scikit-learn
tiktoken
nest-asyncio
asyncio
yfinance
mcp

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
coverage>=7.0.0
mock>=4.0.0